Added ad.html and ad.js. Saved the clicked ads index in sessionstorage and got it in ad.html.

+++ OnlineStore/src/main/java/repositories/AdRepository.java
@@ -14,7 +14,7 @@ import entities.Ad;
 
 //@RepositoryRestResource(path = "ads")
 @RepositoryRestResource
-public interface AdRepository extends CrudRepository<Ad,Integer>{
+public interface AdRepository extends CrudRepository<Ad,Long>{
 	
 	//List<Ad> findByHeadLine(@Param("headline") String headline);
